Fix: Prevents other plugins to tap on done hook
It was using compiler.hooks.done.tapAsync to back slash to replace
backslash with forward-slash. Because the function was not calling the
callback for plugins chain to continue it was preventing other plugins
from getting done hooks event.

As we are only doing synchronous work in that function I used
compiler.hooks.done.tap instead.